# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TemporalLayr Ingestion Server",
    "description": "Production telemetry ingestion dispatch cluster resolving event arrays into backend isolation mechanisms.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/query": {
      "post": {
        "tags": [
          "Observability Dashboard Backend"
        ],
        "summary": "Time Bucket Engine",
        "description": "Time Bucket Engine: executes fast `date_trunc` aggregations natively\nprotecting ORM boundaries by executing raw aggregate counts gracefully!",
        "operationId": "time_bucket_engine_v1_query_post",
        "parameters": [
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DateTruncQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Time Bucket Engine V1 Query Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/search": {
      "post": {
        "tags": [
          "Querying"
        ],
        "summary": "Search Executions",
        "operationId": "search_executions_v1_search_post",
        "parameters": [
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/dashboard": {
      "post": {
        "tags": [
          "Dashboard"
        ],
        "summary": "Create Dashboard",
        "description": "Provisions structural grid bounds encapsulating metrics natively.",
        "operationId": "create_dashboard_v1_dashboard_post",
        "parameters": [
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "List Dashboards",
        "operationId": "list_dashboards_v1_dashboard_get",
        "parameters": [
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DashboardListResponse"
                  },
                  "title": "Response List Dashboards V1 Dashboard Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/stats/top-functions": {
      "get": {
        "tags": [
          "Observability Dashboard Backend"
        ],
        "summary": "Get Top Functions",
        "description": "Top Nodes Aggregator: natively unfolds JSONB arrays exploring heaviest graphs intrinsically!",
        "operationId": "get_top_functions_v1_stats_top_functions_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target enterprise tenant mapping extraction",
              "title": "Tenant Id"
            },
            "description": "Target enterprise tenant mapping extraction"
          },
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Top Functions V1 Stats Top Functions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/errors": {
      "get": {
        "tags": [
          "Observability Dashboard Backend"
        ],
        "summary": "Get Error Rate",
        "description": "Error Rate Indicator: captures anomaly distributions directly bypassing DB heavy load boundaries natively!",
        "operationId": "get_error_rate_v1_stats_errors_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target enterprise tenant mapped globally",
              "title": "Tenant Id"
            },
            "description": "Target enterprise tenant mapped globally"
          },
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Error Rate V1 Stats Errors Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stats/durations": {
      "get": {
        "tags": [
          "Observability Dashboard Backend"
        ],
        "summary": "Get Durations",
        "description": "Latency & Duration Extraction: calculates accurate P95 percentiles\ndynamically in Postgres by parsing node metadata outputs natively!",
        "operationId": "get_durations_v1_stats_durations_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target enterprise tenant mapped globally",
              "title": "Tenant Id"
            },
            "description": "Target enterprise tenant mapped globally"
          },
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Durations V1 Stats Durations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/overview": {
      "get": {
        "tags": [
          "Observability Dashboard Backend"
        ],
        "summary": "Get Overview",
        "description": "Live Data Summary Widget: Returns fast macros avoiding N+1 scaling.\nCalculates moving 1h/24h metrics mapped securely against live telemetry.",
        "operationId": "get_overview_v1_overview_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target enterprise tenant mapping globally",
              "title": "Tenant Id"
            },
            "description": "Target enterprise tenant mapping globally"
          },
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Overview V1 Overview Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/schema": {
      "get": {
        "tags": [
          "Observability Dashboard Backend"
        ],
        "summary": "Get Schema",
        "description": "Auto-Schema Detector: Scans latest N structures flattening deep JSON matrices\nproducing string structural metadata for downstream dashboard queries naturally.",
        "operationId": "get_schema_v1_schema_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target enterprise tenant schema parser",
              "title": "Tenant Id"
            },
            "description": "Target enterprise tenant schema parser"
          },
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Schema V1 Schema Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/dashboard/search": {
      "post": {
        "tags": [
          "Dashboard Ext"
        ],
        "summary": "Search Events",
        "description": "Cursor-bound UI pagination search effectively isolating OFFSET performance drag.",
        "operationId": "search_events_v1_dashboard_search_post",
        "parameters": [
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardDashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dashboard/query": {
      "post": {
        "tags": [
          "Dashboard Ext"
        ],
        "summary": "Query Aggregation",
        "description": "Pipeline aggregation engine seamlessly mapping UI configurations natively into PG.",
        "operationId": "query_aggregation_v1_dashboard_query_post",
        "parameters": [
          {
            "name": "api_key_from_body",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Api Key From Body"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DashboardQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StandardDashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DashboardCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "DashboardCreate"
      },
      "DashboardListResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "created_at"
        ],
        "title": "DashboardListResponse"
      },
...
      "DateTruncQueryRequest": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "from": {
            "type": "string",
            "format": "date-time",
            "title": "From"
          },
          "to": {
            "type": "string",
            "format": "date-time",
            "title": "To"
          },
          "group_by": {
            "type": "string",
            "title": "Group By"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "tenant_id",
          "from",
          "to",
          "group_by",
          "metric"
        ],
        "title": "DateTruncQueryRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}